move : addbutton , removebutton
