+++ src/shared/ui/select/RemoveButton.style.ts
